requirements.txt
openai
pydantic
python-dotenv
streamlit
msal
requests
